+++ space_packet_parser/parseables.py
@@ -33,41 +33,32 @@ class ParsedDataItem:
     long_description: Optional[str] = None
 
 
-@dataclass
-class Packet:
-    """CCSDS Packet
+class RawPacketData(bytes):
+    """A class to represent raw packet data as bytes.
 
-    Can be parsed to populate data items. This ``Packet`` class keeps track
-    of the current parsing position for know where to read from next when
-    parsing data items.
+    This class is a subclass of bytes and is used to represent the raw packet data
+    in a more readable way. It is used to store the raw packet data in the Packet
+    class and used to keep track of the current parsing position.
 
     Parameters
     ----------
     data : bytes
-        The binary data for a single packet.
-    pos : int
-        The bit cursor position in the packet. Default 0.
+        Raw packet data. Full CCSDS packets are always an integer number of bytes.
     """
-    rawdata: bytes
-    pos: Optional[int] = 0
-    parsed_data: Optional[dict] = field(default_factory=lambda: {})
+    def __init__(self, data: bytes, *, pos: int = 0):
+        self.pos = pos
+        self._nbits = len(data) * 8
+        super().__init__()
 
     def __len__(self):
-        """The length of the full packet data object in bits."""
-        return len(self.rawdata) * 8
+        return self._nbits
 
-    @property
-    def header(self):
-        """The header content of the packet."""
-        return dict(list(self.parsed_data.items())[:7])
-
-    @property
-    def data(self):
-        """The user data content of the packet."""
-        return dict(list(self.parsed_data.items())[7:])
+    def __repr__(self):
+        return f"RawPacketData({self}, {len(self)}B, pos={self.pos})"
 
     def read_as_bytes(self, nbits: int) -> bytes:
-        """Read a number of bits from the packet data as bytes.
+        """Read a number of bits from the packet data as bytes. Reads minimum number of complete bytes required to
+        capture `nbits`. Moves `pos` cursor `nbits` forward, even if `nbits` is not an integer number of bytes.
 
         Parameters
         ----------
@@ -83,11 +74,11 @@ class Packet:
             raise ValueError("End of packet reached")
         if self.pos % 8 == 0 and nbits % 8 == 0:
             # If the read is byte-aligned, we can just return the bytes directly
-            data = self.rawdata[self.pos//8:self.pos//8 + nbits // 8]
+            data = self[self.pos//8:self.pos//8 + (nbits+7) // 8]
             self.pos += nbits
             return data
         # We are non-byte aligned, so we need to extract the bits and convert to bytes
-        bytes_as_int = _extract_bits(self.rawdata, self.pos, nbits)
+        bytes_as_int = _extract_bits(self, self.pos, nbits)
         self.pos += nbits
         return int.to_bytes(bytes_as_int, (nbits + 7) // 8, "big")
 
@@ -104,15 +95,49 @@ class Packet:
         : int
             Integer representation of the bits read from the packet
         """
-        int_data = _extract_bits(self.rawdata, self.pos, nbits)
+        int_data = _extract_bits(self, self.pos, nbits)
         self.pos += nbits
         return int_data
 
 
+class CCSDSPacket(dict):
+    """CCSDS Packet
+
+    Container that stores the raw packet data (bytes) as an instance attribute and the parsed
+    data items in a dict interface. A ``CCSDSPacket`` generally begins as an empty dictionary that gets
+    filled as the packet is parsed. The first 7 items in the dictionary make up the
+    packet header (accessed with ``CCSDSPacket.header``), and the rest of the items
+    make up the user data (accessed with ``CCSDSPacket.user_data``). To access the
+    raw bytes of the packet, use the ``CCSDSPacket.raw_data`` attribute.
+
+    Parameters
+    ----------
+    *args : Mapping or Iterable
+        Initial items to store in the packet, passed to the dict() constructor.
+    raw_data : bytes, optional
+        The binary data for a single packet.
+    **kwargs : dict
+        Additional packet items to store, passed to the dict() constructor.
+    """
+    def __init__(self, *args, raw_data: bytes = b"", **kwargs):
+        self.raw_data = RawPacketData(raw_data)
+        super().__init__(*args, **kwargs)
+
+    @property
+    def header(self) -> dict:
+        """The header content of the packet."""
+        return dict(list(self.items())[:7])
+
+    @property
+    def user_data(self) -> dict:
+        """The user data content of the packet."""
+        return dict(list(self.items())[7:])
+
+
 class Parseable(Protocol):
     """Defines an object that can be parsed from packet data."""
-    def parse(self, packet: Packet, **parse_value_kwargs) -> dict:
-        """Parse this entry from the packet data and add the necessary items to the parsed_items dictionary."""
+    def parse(self, packet: CCSDSPacket, **parse_value_kwargs) -> None:
+        """Parse this entry from the packet data and add the necessary items to the packet."""
 
 
 @dataclass
@@ -152,14 +177,13 @@ class SequenceContainer(Parseable):
         self.restriction_criteria = self.restriction_criteria or []
         self.inheritors = self.inheritors or []
 
-    def parse(self, packet: Packet, **parse_value_kwargs) -> dict:
+    def parse(self, packet: CCSDSPacket, **parse_value_kwargs) -> None:
         """Parse the entry list of parameters/containers in the order they are expected in the packet.
 
         This could be recursive if the entry list contains SequenceContainers.
         """
         for entry in self.entry_list:
-            packet.parsed_data = entry.parse(packet=packet, **parse_value_kwargs)
-        return packet.parsed_data
+            entry.parse(packet=packet, **parse_value_kwargs)
 
 
 FlattenedContainer = namedtuple('FlattenedContainer', ['entry_list', 'restrictions'])